+++ src/cohttp_mirage_io.mli
@@ -17,6 +17,8 @@
  * %%NAME%% %%VERSION%%
  *)
 
+(** Cohttp IO implementation using Mirage channels. *)
+
 module Make (Channel: Mirage_channel_lwt.S) : Cohttp.S.IO
   with type 'a t = 'a Lwt.t
    and type ic = Channel.t